LPS-97078 SF - Rename package to internal
